requirements.txt
# Backend
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0

# Frontend будет в папке frontend/
